requirements.txt
streamlit
google-generativeai
edge-tts
asyncio
streamlit-mic-recorder
python-dotenv